third_party/blink/web_tests/virtual/deprecate-unload/http/tests/security/window-events-pass-expected.txt

CONSOLE ERROR: Permissions policy violation: unload is not allowed in this document.
Test that window event handlers set for initial document are kept if security check passes. There will be 2 'PASS' messages below if the test passes:

FAIL
PASS: load fired.
